refactor(dropzone-image): move editing view to own comp, move helpers to reduce prop drilling
