+++ packages/editor/src/plugins/dropzone-image/hooks/use-answer-zones.tsx
@@ -1,24 +1,9 @@
-import { EditorPluginType } from '@editor/types/editor-plugin-type'
 import { useState } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
-import { v4 as uuidv4 } from 'uuid'
 
 import type { DropzoneImageProps } from '..'
 import { AnswerType, type AnswerZoneState } from '../types'
-import {
-  getAnswerZoneImageState,
-  getAnswerZoneText,
-} from '../utils/answer-zone'
-
-export const defaultAnswerZonePosition = {
-  left: 0.05,
-  top: 0.05,
-}
-
-export const defaultAnswerZoneLayout = {
-  width: 0.2,
-  height: 0.1,
-}
+import { duplicateAnswerZone } from '../utils/answer-zone'
 
 export function useAnswerZones(
   answerZones: DropzoneImageProps['state']['answerZones']
@@ -43,47 +28,6 @@ export function useAnswerZones(
     setCurrentAnswerType(type)
   }
 
-  const insertAnswerZone = () => {
-    const currentLength = answerZones.length
-    answerZones.insert(currentLength, {
-      id: `answerZone-${currentLength}`,
-      name: '',
-      position: defaultAnswerZonePosition,
-      layout: defaultAnswerZoneLayout,
-      answers: [],
-    })
-  }
-
-  const duplicateAnswerZone = (idToDuplicate: string) => {
-    const toCopy = answerZones.find((zone) => zone.id.value === idToDuplicate)
-    if (!toCopy) return
-    const currentLength = answerZones.length
-    const newZone = {
-      id: `answerZone-${currentLength}`,
-      name: toCopy.name.value,
-      position: {
-        left: toCopy.position.left.value + defaultAnswerZonePosition.left,
-        top: toCopy.position.top.value + defaultAnswerZonePosition.top,
-      },
-      layout: {
-        width: toCopy.layout.width.value,
-        height: toCopy.layout.height.value,
-      },
-      answers: toCopy.answers.map((answer) => ({
-        id: uuidv4(),
-        image: {
-          plugin: EditorPluginType.Image,
-          state: getAnswerZoneImageState(answer.image.id),
-        },
-        text: {
-          plugin: EditorPluginType.Text,
-          state: getAnswerZoneText(answer.text.id),
-        },
-      })),
-    }
-    answerZones.insert(currentLength, newZone)
-  }
-
   const [answerZoneClipboardItem, setAnswerZoneClipboardItem] =
     useState<AnswerZoneState | null>(null)
 
@@ -104,7 +48,7 @@ export function useAnswerZones(
   useHotkeys(['ctrl+v, meta+v'], (event) => {
     if (!answerZoneClipboardItem) return
     const idToDuplicate = answerZoneClipboardItem.id.value
-    duplicateAnswerZone(idToDuplicate)
+    duplicateAnswerZone(answerZones, idToDuplicate)
     event.preventDefault()
   })
 
@@ -114,7 +58,5 @@ export function useAnswerZones(
     currentAnswerType,
     selectAnswerZone,
     selectCurrentAnswer,
-    insertAnswerZone,
-    duplicateAnswerZone,
   }
 }
